refactor(pipeline): extract result functions to result.go [#466]
Move sendNotification and buildPipelineResult from engine.go to result.go.
